--- client/src/components/All.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function All() {
-  return <div>All</div>;
-}
-
-export default All;
