+++ client/src/components/signupLogin/SignupConfirmed.js
@@ -1,9 +1,11 @@
 import React from "react";
+import { Link } from "react-router-dom";
 
 function SignupConfirmed() {
   return (
     <>
       <div>SignupConfirmed</div>
+      <Link to="/login">Login</Link>
     </>
   );
 }
